backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
langgraph>=0.0.1
langchain>=0.1.0
langchain-openai>=0.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
urllib3
apscheduler>=3.10.1
httpx>=0.24.1
markdown>=3.4.0
